#28 moved up WebEntityConstants and fixed unit tests
