+++ entity-web/src/main/java/eu/europeana/entity/web/jsonld/ConceptSetSerializer.java
@@ -6,7 +6,7 @@ import org.apache.stanbol.commons.jsonld.JsonLdPropertyValue;
 import org.apache.stanbol.commons.jsonld.JsonLdResource;
 
 import eu.europeana.entity.definitions.model.search.result.ResultSet;
-import eu.europeana.entity.web.controller.WebEntityConstants;
+import eu.europeana.entity.definitions.vocabulary.WebEntityConstants;
 import eu.europeana.entity.web.model.view.ConceptView;
