+++ entity-web/src/main/java/eu/europeana/entity/web/jsonld/SuggestionSetSerializer.java
@@ -13,7 +13,7 @@ import org.apache.stanbol.commons.jsonld.JsonLdResource;
 import eu.europeana.entity.definitions.model.ResourcePreview;
 import eu.europeana.entity.definitions.model.search.result.ResultSet;
 import eu.europeana.entity.definitions.model.vocabulary.EntityTypes;
-import eu.europeana.entity.web.controller.WebEntityConstants;
+import eu.europeana.entity.definitions.vocabulary.WebEntityConstants;
 import eu.europeana.entity.web.model.view.AgentPreview;
 import eu.europeana.entity.web.model.view.EntityPreview;
 import eu.europeana.entity.web.model.view.PlacePreview;
